Cambiar estado de inmueble y Agregar nuevo inmueble. OK

+++ app/src/main/java/com/example/inmobiliarialabiii/model/Inmueble.java
@@ -8,7 +8,7 @@ public class Inmueble implements Serializable {
     private String uso;
     private String tipo;
     private int ambientes;
-    private double superficie;
+    private int superficie;
     private double latitud;
     private double valor;
     private String imagen;
@@ -17,7 +17,7 @@ public class Inmueble implements Serializable {
     private int idPropietario;
     private Propietario duenio;
 
-    public Inmueble(int idInmueble, String direccion, String uso, String tipo, int ambientes, double superficie, double latitud, double valor, String imagen, boolean disponible, double longitud, int idPropietario, Propietario duenio) {
+    public Inmueble(int idInmueble, String direccion, String uso, String tipo, int ambientes, int superficie, double latitud, double valor, String imagen, boolean disponible, double longitud, int idPropietario, Propietario duenio) {
         this.idInmueble = idInmueble;
         this.direccion = direccion;
         this.uso = uso;
@@ -80,7 +80,7 @@ public class Inmueble implements Serializable {
         return superficie;
     }
 
-    public void setSuperficie(double superficie) {
+    public void setSuperficie(int superficie) {
         this.superficie = superficie;
     }
